Remove old attacker_id key, move AdvCP test within AdvCP directory
