+++ opencda/scenario_testing/config_yaml/advcp/spoofing_forward.yaml
@@ -1,6 +1,6 @@
 # For 2cars_advcp_attack_check scenario
 mode: spoof
-attacker_id: cav-2
+attacker_ids: [cav-2]
 density: 2 # 0=replace, 1=dense_a, 2=dense_all, 3=sampled
 dense_distance: 10.0 # used by dense_a and dense_all early spoofing
 default_size: [4.5, 2.0, 1.6]
